requirements.txt
# Core dependencies
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.0

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.4
pyxpdf==0.3.7

# Vector database
chromadb==0.4.22

# LangChain and related packages
langchain==0.1.4
langchain-openai==0.0.5
langchain-community==0.3.21
langchain-chroma==0.2.2

# PubMed integration
biopython==1.83
metapub==0.5.5
pubmed2pdf==0.0.7
entrezpy==2.1.3

# Utilities
requests==2.31.0
beautifulsoup4==4.12.3
tqdm==4.66.2
pycryptodome==3.19.0
pandas==2.1.4
numpy==1.26.4
matplotlib==3.8.3
seaborn==0.13.1

# Web scraping and parsing
lxml==5.1.0
html5lib==1.1
soup2dict==1.0.4

# Additional dependencies for deployment
protobuf>=4.25.1
typing-extensions>=4.5.0
altair>=4.0
tenacity>=8.2.0
pydeck>=0.8.0
watchdog>=3.0.0

# Date and time handling
python-dateutil==2.8.2
pytz==2024.1

# File handling and compression
pyarrow>=14.0.1
zipfile38>=0.0.3
